refactor: api 폴더구조 적용
