+++ frontend/src/api/createBook.ts
@@ -1,0 +1,13 @@
+import { AxiosResponse } from 'axios';
+import { axiosInstance } from '.';
+
+interface bookData {
+	title: string;
+	authors: string[];
+	publisher: string;
+}
+
+export const getBooksByTitle = (
+	searchText: string,
+): Promise<AxiosResponse<bookData[]>> =>
+	axiosInstance.get(`/books/bookInfo?bookTitle=${searchText}`);
