+++ frontend/src/api/hooks/createBooks/useGetBooksByTitle.tsx
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+import { getBooksByTitle } from '../../createBook';
+
+const useGetBooksByTitle = (searchText: string) =>
+	useQuery(['books'], () => getBooksByTitle(searchText), {
+		refetchOnWindowFocus: false,
+		enabled: false,
+	});
+
+export default useGetBooksByTitle;
