+++ frontend/src/components/BooksCreate/ModalForTitle.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import styled from 'styled-components';
-import useGetBooksByTitle from '../../api/hooks/useGetBooksByTitle';
+import useGetBooksByTitle from '../../api/hooks/createBooks/useGetBooksByTitle';
 import { useAppDispatch } from '../../redux/hooks';
 import notify from '../../utils/notify';
 import Button from '../common/Button';
